refactor: MyTeamCard 컴포넌트 정리
